frontend/requirements.txt
streamlit==1.28.0
streamlit-chat==0.1.1
requests==2.31.0
sseclient-py==1.7.2 